refactor(homepage): split card and cards components
